backend/pyproject.toml
[project]
name = "codegambit-backend"
version = "0.1.0"
description = "CodeGambit - AI-powered developer training platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0,<3.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36,<3.0",
    "aiosqlite>=0.20.0",
    "anthropic>=0.42.0,<1.0",
    "docker>=7.1.0",
    "pyyaml>=6.0.2",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
